feat: data bindings
Switched from HTMX to Alpine.js for data bindings. The only thing we
needed from HTMX was the AJAX attributes. We had to use extensions for
templating, and we could not bind data to the template. Alpine.js lets
us reuse the DOM and bind data to it.

With the JSON bound to an `<textarea>`, and individual inputs bound to
the properties of the JSON, changes to the inputs are reflected in the
source JSON.

For fetching the JSON, we set up a `request` setter that menu action
would use, and use normal `fetch` to load the data.

The modal watches for changes in the data and appears.

To facilitate the reuse of templates, an `x-component` attribute is
added to be used on a `<template>` element. These templates are then
converted into a custom element.